add initial mknImage with the crc32 command
Bring in clap, add a nimage::util mod, and do some stuff with
a CmdHandler trait to build up a command table. Function pointers are
weird in Rust and I couldn't get a boxed FnOnce to work, so use a strait
and nil structs for now.

Also disable non_snake_case warnings so that cargo doesn't complain
about mknImage not being snake case. Apparently the binary name is
treated as a crate name, which is annoying.