util: add SliceReader, a wrapper for reading from u8 slices
A wrapper for a borrowed u8 slice that keeps track of the current
position. The standard library implements Read for &[u8] but that method
updates the slice to remove the read out bytes, whereas SliceReader
never mutates the original slice.

Also includes methods to read u8, u32le, and u64le integers out of the
slice, handling the tricky use of from_le_bytes, and implements the Seek
trait so we can skip around.