format: save the read crc in ImageHeader
The "mknImage check" command should print the header CRC, and it should
print the actual CRC read frm the input file (which could be different
than a recalculated CRC if the input file had nonzero bytes in unused
regions).

To support this, store the CRC in the ImageHeader struct, nothing that
it's only valid immediately calling from_bytes or write_to.