util: add human_size(_extended) functions, impl Display for Input
The human_size functions are to pretty-print file sizes.

Modify the Input enum so that the File variant includes a copy of the
filename that can be used when printing. This is because the File object
itself doesn't remember the name it was opened with (with links and
relative paths, any number of path names could point to the same file).