format: implement TryFrom<&str> for PartType
Will be used during mknImage create. To avoid duplicating lists of the
type->name mapping, add a static array and search through it. Sure it's
an O(n) lookup but n==10 so idgaf.

Also in check.rs move the clippy allow comparison_chain attribute to the
function instead of the if statement. Cargo fmt (but not clippy???)
complains that attributes aren't supported on if statements.