version 0.2.0-dev
v0.1.0 has been tagged, with mknImage create/check support for the
nImage header version 2.

Start development on nImage version 3, which will be easier to implement
swdl and auto-compression for. nimage-rs will not be backwards
compatible with v2 images because I don't want to maintain that.

Planned nImage format v3 changes:
  * remove compression-specific part types, shrink the list down to
    BootImg, BootTar, Rootfs, and RootfsRw.
  * Add a compression type field to the part header.
  * Only two supported compression formats: zstd (natively supported),
    and libarchive (anything that bsdcat(1) can decompress through
    a pipe, and will require bsdcat to be installed for swdl)