add xxhio: Reader/Writer wrappers for twox_hash::XxHash32
I've decided that in nImage v3, CRC32 will be replaced by xxHash32,
since CRC32 is slow slow that it's noticeable in un-optimized builds.
Based on some benchmarking, xxHash32 is about 7x faster than CRC32 in
Rust release builds.

Use a forked twox-hash repo for now, where I added a direct method to
get the 32-bit hash and, more importantly, store total_len as a u64 on
XxHash32 rather than a u32, which is needed to support parts larger than
4GiB.

I'll send a PR to twox_hash so that hopefully I can switch back to using
official versions of that module. For now I'm happy that Cargo makes it
easy to set custom git repos for dependencies.