v3 support: update mknImage create to include compression mode
Change the part argument format from TYPE:FILE to FILE:TYPE[:COMP] so
that shell completion can be used for FILE, and the optional COMP is at
the end which is more logical.

A side effect of this is that FILE can't contain any colons, since
they'll be interpreted as field separators, but that's not a bad
limitation for a cleaner implementation.