cargo: use pkg-config for zstd
Linking against the system libzstd library significantly speeds up
builds.

On zstd-rs master the pkg-config feature is forwarded from zstd-sys to
zstd, but that's not released yet. Instead add an explicit zstd-sys
dependency to turn on that feauture. Use a wildcard version since zstd
and zstd-safe depend on exact versions of zstd-safe and zstd-sys,
respectively.