rework zstd-sys/pkg-config dependency
When cross-compiling, I don't want pkg-config enabled for zstd-sys
because I don't have a system version of libzstd installed for cross
targets.

One would think that it'd be easy to handle this with
    [target.x86_64-unknown-linux-gnu.dependencies]
    zstd-sys = { version = "*", features = ["pkg-config"] }
so that pkg-config is only turned on for x86_64 builds. Unfortunately,
Cargo uses the native target for build.rs scripts, and forwards features
there accordingly. Thus, zstd-sys' build.rs is compiled with pkg-config,
and then it complains that it can't find libzstd.

Furthermore, Cargo doesn't have any support for changing the default
features based on target, as nice as that would be.

Hack around this mess by by adding a pkg-config feature to nimage-rs
that's enabled by default. Cross builds will need to compile with
--no-default-features to turn this off.