swdl: actually program raw parts to disk
Add some functions to get the destination path for raw parts, and use
that. Some cleanup is needed for using /dev/null on x86 builds more
cleanly, but it works for now.

Open the output file descriptor with O_SYNC (using libc and
OpenOptionsExt) so that the progress bar tracks actual progress,
sacrificing a bit of performance. It's worth it to not have the progress
bar zip by followed by a long delay while the file descriptor is closed.
Decompressing parts seems to get really slow towards the end, need to
see if the zstd decompressor is doing weird stuff I guess.